i was happy with the results of the stablizsr so i turend my attension to creat a real sensors, and by real i mean fake, webots fake. but it will be more reall becouse its based on GPS and a compass.
i added a compass and removed a stupad interaction betwin the webots classes.
 added a GPS and implemented them.

changed the name of the stupadly long and incorrect name of GlobaleOriantaionSensor to just LFS to say its an Lframe sensor.

implemented a class of GPS_COM_LFS
that dose just that

the sensor is garbage, it needs to know which direction is down... maybe incorporate the IMU or an accelerometer for that.

but the direction the project is going is great... maybe i will start implementing the Navegator state chart soon. its not suposed to be a hassel but its on the bucket list.

after the sensors i will continue to obsticalls avoidance behavior, and navagator behavior.

the IFfame orientation has bean configured bt its origen may very.
the y axis is the north, the -z axis is gravity.

+++ BehaviorLFrameMaker.cpp
@@ -14,13 +14,13 @@ Frame BehaviorLFrameMaker::calcLframe()
     mtx.unlock();
     auto now = system_clock::now();
 #ifdef WEBOTS_STEP_TIME_MS
-    float dt = WEBOTS_STEP_TIME_MS;
+    float dt = (float)WEBOTS_STEP_TIME_MS;
 #else
     auto dt = duration_cast<milliseconds>(now - last).count(); 
 #endif // WEBOTS_STEP_TIME_MS
 
     last = now;
-    lFrame.pos += (vel.block<3, 1>(0, 0)).cast<double>() * dt;
+    lFrame.pos += vel.block<3, 1>(0, 0) * dt;
     lFrame.rotate(vel[3] * dt);
     std::cout << lFrame << std::endl;
     return lFrame;
