+++ Frame.h
@@ -3,21 +3,21 @@
 #include <Eigen/dense>
 #include <iostream>
 
-using Eigen::Vector3d;
-using Eigen::Matrix2d;
+using Eigen::Vector3f;
+using Eigen::Matrix2f;
 
 struct Frame {
-	Vector3d pos;
-	Matrix2d ori;
+	Vector3f pos;
+	Matrix2f ori;
 	Frame();
-	Frame(Vector3d pos, double angle);
+	Frame(Vector3f pos, double angle);
 
 	bool operator==(const Frame& other) const
 	{
 		return pos == other.pos && ori == other.ori;
 	}
 
-	Frame(const Vector3d& pos, const Matrix2d& ori)
+	Frame(const Vector3f& pos, const Matrix2f& ori)
 		: pos(pos), ori(ori)
 	{
 	}
