+++ PIDw.cpp
@@ -9,7 +9,7 @@ PIDw::PIDw(float p, float i, float d)
     lastTime(std::chrono::system_clock::now()) {
 }
 
-float PIDw::update(const Matrix2d& setpoint, const Matrix2d& measured_value)
+float PIDw::update(const Matrix2f& setpoint, const Matrix2f& measured_value)
 {
     // Get the current time and calculate the time difference (dt) in milliseconds
     auto now = std::chrono::system_clock::now();
@@ -25,8 +25,8 @@ float PIDw::update(const Matrix2d& setpoint, const Matrix2d& measured_value)
 
     // Update the last time to the current time
     lastTime = now;
-    Vector2d a = setpoint.row(0);
-    Vector2d b = measured_value.row(0);
+    Vector2f a = setpoint.row(0);
+    Vector2f b = measured_value.row(0);
     // Calculate the error
     if(a[X] * b[Y] > b[X] * a[Y])
         error = std::acos(a.dot(b));
