+++ PIDw.h
@@ -5,12 +5,12 @@
 #include <chrono>
 #include "Primitives.h"
 
-class PIDw : public Controller_I<float, const Matrix2d&> {
+class PIDw : public Controller_I<float, const Matrix2f&> {
 	float kp, ki, kd;
 	float error, prev_error, integral;
 	float output;
 	std::chrono::system_clock::time_point lastTime;
 public:
 	PIDw(float p, float i, float d);
-	float update(const Matrix2d& setpoint, const Matrix2d& measured_value) override;
+	float update(const Matrix2f& setpoint, const Matrix2f& measured_value) override;
 };
